+++ internal/simd/simd_vec_neon.h
@@ -214,18 +214,18 @@ template <> class simd_vec<float, 4> {
 
         alignas(16) float comp[4];
         vst1q_f32(comp, vec_);
-        UNROLLED_FOR(i, 4, { comp[i] = std::sqrt(comp[i]); })
+        UNROLLED_FOR(i, 4, { comp[i] = sqrtf(comp[i]); })
         return simd_vec<float, 4>{comp, simd_mem_aligned};
     }
 
     force_inline simd_vec<float, 4> log() const {
         alignas(16) float comp[4];
         vst1q_f32(comp, vec_);
-        UNROLLED_FOR(i, 4, { comp[i] = std::log(comp[i]); })
+        UNROLLED_FOR(i, 4, { comp[i] = logf(comp[i]); })
         return simd_vec<float, 4>{comp, simd_mem_aligned};
     }
 
-    force_inline float length() const { return std::sqrt(length2()); }
+    force_inline float length() const { return sqrtf(length2()); }
 
     force_inline float length2() const {
         alignas(16) float comp[4];
@@ -444,7 +444,7 @@ template <> class simd_vec<float, 4> {
         alignas(16) float comp1[4], comp2[4];
         vst1q_f32(comp1, v1.vec_);
         vst1q_f32(comp2, v2.vec_);
-        UNROLLED_FOR(i, 4, { comp1[i] = std::pow(comp1[i], comp2[i]); })
+        UNROLLED_FOR(i, 4, { comp1[i] = powf(comp1[i], comp2[i]); })
         return simd_vec<float, 4>{comp1, simd_mem_aligned};
     }
